add command-line arguments interface
